backend/requirements.txt
bcrypt==4.2.0
ecdsa==0.19.0
greenlet==3.1.1
passlib==1.7.4
pyasn1==0.6.1
PyJWT==2.9.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
